Raise if WOLoginItem initialized with a nil path
Initialization is one of the "edge" boundaries of an API where we want
to be strict about what we will accept in order to detect errors sooner.

I just ran into a crash caused by a build bug (missing file) which in
turn caused a nil value to be passed into the WOLoginItem initialization
codepath; this parameter check would have helped me discover the cause
of the problem a little sooner.

Signed-off-by: Wincent Colaiuta <win@wincent.com>