Deprecate the createDirectoryAtPath:attributes:createParents: method
This method is made redundant by the createDirectoryAtPath:
withIntermediateDirectories:attributes:error: method which as added to
NSFileManager in Leopard, so turn it into a simple wrapper for that
method and mark it as deprecated.

This fixes deprecation warnings building against the 10.6 SDK (not
tested against the 10.5 SDK) about the use of createDirectoryAtPath:
attributes: method which is currently deprecated.

Signed-off-by: Wincent Colaiuta <win@wincent.com>