Fix bug in error-path of pathForFolder:domain:create:
This bug was discovered while running the unit tests on a new system
with a clean Snow Leopard install. Previously the error-path containing
the bug had never been exercised.

Flesh out the unit tests to increase the likelihood of such a bug being
caught in the future if it ever returns (it is not possible to
guarantee 100% that the error-path can be triggered, because we don't
know what directories will exist on the system where the tests are
running).

Signed-off-by: Wincent Colaiuta <win@wincent.com>