Add benchmarks tool
Add tool for running some simple benchmarks. The first benchmark added
is a comparison of the speed of using the new -[NSArray map:] method
with a couple of variants of explicit manual enumeration.

     -[NSArray map:]: 3.451480/0.092237/3.543717 (user/system/total)
  manual enumeration: 3.562207/0.093985/3.656192 (user/system/total)
    fast enumeration: 3.479888/0.124764/3.604652 (user/system/total)

The results should be taken with a grain of salt as multiple runs can
produce a both relative (changes of order) and absolute variation, but
at least at first glance it would appear that for non
performance-critical applications the election of method really isn't
that important.

Signed-off-by: Wincent Colaiuta <win@wincent.com>