Use explicit self in WOLoginItemList init
This has always been best practice in `init` methods, but I have been
inconsistent in following it. This SO post explains the difference:

http://stackoverflow.com/a/19665933/1626737

Effectively, after `self = [super init]`, the explicit "self" referred
to in `self->listRef` may not be the same as the implicit "self" in a
naked `listRef` expression.

People have been reporting crash on launch:

- issue: https://wincent.com/issues/2173
- see also comments here: http://www.macupdate.com/app/mac/10042/synergy

I am not sure if this will fix that issue, but it might, so it seems
worth giving it a try. If it does fix it, I should audit all init
methods in the code base and make sure they do the right thing.

Signed-off-by: Greg Hurrell <greg@hurrell.net>